Shrink Timing struct from 16 to 8 bytes (#123)
* Shrink Timing struct from 16 to 8 bytes

Use u32 for delay_millis (supports ~49 days) and f32 for value,
halving per-record memory during long benchmark runs.

Made-with: Cursor

* Use f32 in SearcherResults

No need to widen to f64 for JSON serialization.

Made-with: Cursor

---------

+++ src/save_jsonl.rs
@@ -16,12 +16,12 @@ pub fn save_timings_as_jsonl(
         if absolute_time {
             record.insert(
                 "timestamp".to_string(),
-                start_timestamp_millis + timing.delay_millis,
+                start_timestamp_millis + f64::from(timing.delay_millis),
             );
         } else {
-            record.insert("delay".to_string(), timing.delay_millis);
+            record.insert("delay".to_string(), f64::from(timing.delay_millis));
         }
-        record.insert(value_name.to_string(), timing.value);
+        record.insert(value_name.to_string(), f64::from(timing.value));
         data.push(record);
     }
